allowance for non-docker services

+++ initdb.sh
@@ -8,7 +8,8 @@ function usage {
 `basename "$0"`: [options]
 Options:
   -h    this help message
-  -r    reset postgres docker data volume for this project
+  -D    run services in docker (docker-compose-services)
+  -r    reset postgres data
   -R    reset all docker data volumes for this project
   -b    (re)build the app docker container
   -g    update a local git repository (initialise if nencessary)
@@ -20,9 +21,10 @@ USAGE
 dc_reset=0
 dc_reset_all=0
 dc_build_app=1
+dc_docker=0
 git_init=0
 
-args=`getopt hgbrR $*` || { usage && exit 2; }
+args=`getopt hgdbrR $*` || { usage && exit 2; }
 set -- $args
 for opt
 do
@@ -31,6 +33,10 @@ do
       usage
       exit 1
       ;;
+    -D)
+      dc_docker=1
+      shift
+      ;;
     -b)
       dc_build_app=1
       shift
@@ -38,6 +44,7 @@ do
     -R)
       dc_reset_all=1
       dc_reset=1
+      dc_docker=1
       shift
       ;;
     -r)
@@ -52,32 +59,43 @@ do
 done
 
 if [ ${dc_reset} != 0 ]; then
-  # check to see if anything is running
-  if [ ! -z "`docker ps -q -f name=${COMPOSE_PROJECT_NAME}`" ]; then
-    echo '*>' Stopping all running containers
-    docker-compose down
-  fi
-  if [ ${dc_reset_all} != 0 ]; then
-    # remove all project volumes
-    volumes=`docker volume ls -q -f name=${COMPOSE_PROJECT_NAME}`
+  if [ ${dc_docker} != 0 ]; then
+    # check to see if anything is running
+    if [ ! -z "`docker ps -q -f name=${COMPOSE_PROJECT_NAME}`" ]; then
+      echo '*>' Stopping all running containers
+      docker-compose down
+    fi
+    if [ ${dc_reset_all} != 0 ]; then
+      # remove all project volumes
+      volumes=`docker volume ls -q -f name=${COMPOSE_PROJECT_NAME}`
+    else
+      # remove project postgres volume
+      volumes=`docker volume ls -q -f name=${COMPOSE_PROJECT_NAME}_data-pgdata`
+    fi
+    if [ ! -z "${volumes}" ]; then
+      echo '*>' Removing volumes ${volumes}
+      docker volume rm ${volumes}
+    fi
   else
-    # remove project postgres volume
-    volumes=`docker volume ls -q -f name=${COMPOSE_PROJECT_NAME}_data-pgdata`
-  fi
-  if [ ! -z "${volumes}" ]; then
-    echo '*>' Removing volumes ${volumes}
-    docker volume rm ${volumes}
+    PGPASSWORD="${POSTGRES_PASSWORD}" psql -h ${DBHOST} -p ${DBPORT} postgres postgres <<SQL
+drop database ${DBNAME};
+drop user ${DBUSER};
+drop role ${DBROLE};
+SQL
   fi
 fi
 
-if [ -z "`docker ps -q -f name=${COMPOSE_PROJECT_NAME}`" ]; then
-  echo '*>' Starting ${COMPOSE_PROJECT_NAME} service containers
-  docker-compose -f docker-compose-services.yml up -d
-  echo '*>' Waiting a few seconds for the database to come up
-  sleep 10.0
+if [ ${dc_docker} != 0 ]; then
+  if [ -z "`docker ps -q -f name=${COMPOSE_PROJECT_NAME}`" ]; then
+    echo '*>' Starting ${COMPOSE_PROJECT_NAME} service containers
+    docker-compose -f docker-compose-services.yml up -d
+    echo '*>' Waiting a few seconds for the database to come up
+    sleep 10.0
+  fi
 fi
 
 # create a role with a user, use permission inheritance for convenience
+echo "Setting up database roles"
 PGPASSWORD="${POSTGRES_PASSWORD}" psql -h ${DBHOST} -p ${DBPORT} postgres postgres <<SQL
 create role ${DBROLE} createdb;
 create user ${DBUSER} createrole inherit password '${DBPASS}';
@@ -88,14 +106,11 @@ alter role ${DBROLE} set timezone to 'UTC';
 SQL
 
 # create the database(es)
-for db_name in ${DBNAME}
-do
-  echo "Creating database: ${db_name}"
-  PGPASSWORD="${POSTGRES_PASSWORD}" psql -h ${DBHOST} -p ${DBPORT} postgres postgres <<SQL
-create database ${db_name} with owner ${DBROLE};
-grant all privileges on database ${db_name} to ${DBROLE};
+echo "Creating database: ${DBNAME}"
+PGPASSWORD="${POSTGRES_PASSWORD}" psql -h ${DBHOST} -p ${DBPORT} postgres postgres <<SQL
+create database ${DBNAME} with owner ${DBROLE};
+grant all privileges on database ${DBNAME} to ${DBROLE};
 SQL
-done
 
 # do intialisation in sub-shell
 (
